refactor: restructure project with proper package organization
- Move main.py logic to src/main.py and src/tools.py
- Update pyproject.toml with proper build configuration and entry point
- Add src/__init__.py for proper Python package structure
- Update test_client.py to use new module path
- Remove old main.py file from root directory
