+++ src/__init__.py
@@ -1,0 +1,3 @@
+"""MCP server for Weaviate vector database operations."""
+
+__version__ = "0.1.0"
